Add the basic SQLite statement logic
This is a working prototype, kinda. Trying it with a real database, I
got a `"LIMIT ?;" | -d 3` test working properly, but the -t flag
wasn't giving me anything. Maybe it's the query, but it definitely
needs better error messages. Other things to do include actually using
the custom separators (maybe by passing them with open_mode in an
options struct), blob parameters (-b and -F could be good flags for
blobs as arguments and files, respectively), custom strings for NULL
and BLOB results (-N and -B flags? also, a blob could be represented
by a string with non-printable bytes represented with \xXX or
something else), and perhaps split the SQLite part of the program into
its own file. Oh, and it doesn't handle multiple statements well so
far, I think. I really need to test this thing thoroughly. And to
write documentation, since -h is not a substitute.