Create a POSIX compliant Makefile
Tested on both GNU and NetBSD Make. Additionally, created the targets
"install" and "clean".

Too bad I can't concatenate a string to an existing macro with
something like A = $(A) B, since it throws a "recursive macro" error
on BSD, but whatever, it's not like I can't create another macro.