Add custom output strings for BLOB and NULL values
This commits two new switches: -N and -B. By default, there's no
string that substitutes a BLOB value, outputing the bytes instead by
escaping the non-printable characters. The usual sequences are used
for escaping, fallbacking to \xNN sequences for the rest of the
characters. There might be a bit of memory overuse, allocating 4 times
the length of the BLOB value to store the escape strings, but it's
freed quickly.

Probably the next step is to add a parameter to use files as
BLOBS.